backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
pymysql==1.1.1
httpx==0.28.1
python-multipart==0.0.20
pydantic-settings==2.7.0
